+++ RealUI_Skins/AddOns/Ace3.lua
@@ -24,13 +24,13 @@ local function SkinAceConfig()
         popup.accept:SetPushedTexture("")
         popup.accept:SetHighlightTexture("")
         Base.SetBackdrop(popup.accept, Color.button)
-        Base.SetHighlight(popup.accept, "backdrop")
+        Base.SetHighlight(popup.accept)
 
         popup.cancel:SetNormalTexture("")
         popup.cancel:SetPushedTexture("")
         popup.cancel:SetHighlightTexture("")
         Base.SetBackdrop(popup.cancel, Color.button)
-        Base.SetHighlight(popup.cancel, "backdrop")
+        Base.SetHighlight(popup.cancel)
     end
 end
 local function SkinAceGUI()
@@ -51,7 +51,7 @@ local function SkinAceGUI()
             Base.SetBackdrop(widget.border, frameColor)
         end,
         Frame = function(widget)
-            Base.SetBackdrop(widget.frame)
+            Skin.FrameTypeFrame(widget.frame)
 
             -- Regions
             local regions = {widget.frame:GetRegions()}
@@ -119,7 +119,7 @@ local function SkinAceGUI()
         end,
         Window = function(widget)
             -- /run LibStub("AceGUI-3.0"):Create("Window")
-            Base.SetBackdrop(widget.frame)
+            Skin.FrameTypeFrame(widget.frame)
 
             -- Regions
             local regions = {widget.frame:GetRegions()}
@@ -243,7 +243,7 @@ local function SkinAceGUI()
                 bg:SetScript("OnMouseDown", widget.frame:GetScript("OnMouseDown"))
                 bg:SetScript("OnMouseUp", widget.frame:GetScript("OnMouseUp"))
                 Base.SetBackdrop(bg, Color.button, 0.3)
-                Base.SetHighlight(bg, "backdrop")
+                Base.SetHighlight(bg)
 
                 widget.checkbg:SetTexture("")
                 widget.highlight:SetTexture("")
@@ -264,7 +264,9 @@ local function SkinAceGUI()
                     self.g = g
                     self.b = b
                     self.a = a or 1
-                    bg:SetBackdropColor(r, g, b, a)
+                    if r then
+                        bg:SetBackdropColor(r, g, b, a)
+                    end
                 end
                 function widget:SetDisabled(disabled)
                     self.disabled = disabled
@@ -317,7 +319,7 @@ local function SkinAceGUI()
                 widgets["Dropdown-Item"](widget)
             end,
             ["Dropdown-Pullout"] = function(widget)
-                Base.SetBackdrop(widget.frame)
+                Skin.FrameTypeFrame(widget.frame)
 
                 local scrollFrame = widget.scrollFrame
                 scrollFrame:SetPoint("TOPLEFT", widget.frame, "TOPLEFT", 1, -12)
@@ -417,15 +419,15 @@ local function SkinAceGUI()
                 local disabled = button:GetDisabledTexture()
                 disabled:SetColorTexture(0, 0, 0, .3)
                 disabled:SetDrawLayer("OVERLAY")
-                Base.SetBackdrop(button, Aurora.Color.button)
 
                 local arrow = button:CreateTexture(nil, "ARTWORK")
                 arrow:SetPoint("TOPLEFT", 4, -7)
                 arrow:SetPoint("BOTTOMRIGHT", -5, 6)
                 Base.SetTexture(arrow, "arrowDown")
+                button._auroraTextures = {arrow}
 
-                button._auroraHighlight = {arrow}
-                Base.SetHighlight(button, "texture")
+                Base.SetBackdrop(button, Aurora.Color.button)
+                Base.SetHighlight(button)
 
                 local bg = _G.CreateFrame("Frame", nil, frame)
                 bg:SetPoint("BOTTOMRIGHT", button, "BOTTOMLEFT", 1, 0)