+++ nibRealUI/Infobar/Blocks.lua
@@ -1581,7 +1581,17 @@ function Infobar:CreateBlocks()
                 return _G.ARTIFACT_BAR_COLOR:GetRGB()
             end,
             IsValid = function(Art)
-                return C_AzeriteItem.HasActiveAzeriteItem() and not C_AzeriteItem.IsAzeriteItemAtMaxLevel()
+                local isMaxLevel = C_AzeriteItem.IsAzeriteItemAtMaxLevel();
+                if isMaxLevel then
+                    return false
+                end
+
+                azeriteItemLocation = C_AzeriteItem.FindActiveAzeriteItem()
+                if RealUI.isPatch then
+                    return azeriteItemLocation and azeriteItemLocation:IsEquipmentSlot() and C_AzeriteItem.IsAzeriteItemEnabled(azeriteItemLocation)
+                else
+                    return azeriteItemLocation and azeriteItemLocation:IsEquipmentSlot()
+                end
             end,
             SetTooltip = function(Art, tooltip)
                 local xp, totalLevelXP, name, currentLevel = Art:GetStats()